+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Change History
 
+## master
+- Updated status bar to be more readable
+
 ## 0.2.1
 - Directory list is now windowed, so browsing larger directories is feasible.